refactor(game) : continueGame 위치변경
